requirements.txt
fastapi[standard]>=0.110.0
sqlmodel>=0.0.16
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.9
boto3>=1.34.0
psycopg2-binary>=2.9.9
supabase>=2.4.0
python-dotenv>=1.0.1
slowapi>=0.1.9
filetype>=1.2.0
